refactor: Facade 계층 제거 (#447)
